requirements.txt
fastapi
gunicorn
uvicorn
databases[postgresql]
SQLAlchemy
python-dotenv
pydantic[email]
pydantic
httpx
boto3
aiofiles
beautifulsoup4
lxml
passlib
bcrypt
python-jose[pycryptodome]
psycopg2-binary
asyncpg
python-multipart
apns2 